IAD-490(Update the workflow when moving email)
Description
We need to update the way the email is displayed after doing an action.
Today, a banner is displayed and the email details is still displayed to user.

To do
When user moves an email to:
archive
spam
trash
unread
then the user is redirected to the list of emails
and a banner is displayed to confirm the action is done.